--- a/theano/scan_module/scan_op.py
+++ b/theano/scan_module/scan_op.py
@@ -1936,7 +1936,7 @@ class Scan(PureOp):
                 dc_dxts_idx += 1
             else:
                 if isinstance(dC_douts[i].type, DisconnectedType):
-                    c_dxts_idx += 1
+                    dc_dxts_idx += 1
                     continue
                 else:
                     if diff_outputs[i] in known_grads: