NOTE : I'm still in the process of testing this change but it might still be good to start getting feedback. Since this changes the graphs that result from taking gradients of Scan nodes, it may mean that new optimizations that were previously not needed would now be more useful/important.

This PR modifies the way Scan takes the gradient of it's inner function. Instead of doing nb_inputs * nb_outputs calls to grad to get the gradient of its inner function, it now does only one such call. This makes grad() generally faster but even more so in the case of taking the gradient of nested scans in which the number of calls to grad() would increase exponentially with the number of levels of nested scans. Because of the reduced number of calls to grad() the resulting scan grads have simple graphs and less duplicate nodes so optimization is faster and (because we don't have optimizations for every single case) execution is sometimes faster. Also, execution with no/few optimizations (i.e. FAST_COMPILE), execution is much much faster.

On the LSTM tutorial, on my machine, the time to take the gradients goes from 0.30 to 0.15 seconds, the compilation time goes from 16s to 13s and the time to run the scan_grad drops by ~40% resulting in a global ~20% drop in execution time.

On the following (not-rigorous) graph using nested scans, the time to take the gradients goes from 6.9s to 0.7s seconds, the compilation time goes from 18s to 7s but the execution time stayed the same because the resulting Theano function was the same as without this change.
```
def test_grad_speed():
    a = tensor.fmatrix()
    b = tensor.fmatrix()
    c = tensor.fmatrix()
    d = tensor.fmatrix()
    e = tensor.fmatrix()

    def outer_step(a_1, b_1, c_1, d_1, e_1):
        def inner_step(a_2, b_2, c_2, d_2, e_2):
            total = a_2 + b_2 + c_2 + d_2 + e_2
            total = total + 1 + 1 + 1 + 1
            return total, total, total, total, total

        result_inner, _ = theano.scan(fn=inner_step,
                                      sequences=[a_1, b_1, c_1, d_1, e_1],
                                      name="inner_scan")

        total = (result_inner[0][-1] + result_inner[1][-1] +
                 result_inner[2][-1] + result_inner[3][-1] +
                 result_inner[4][-1])
        return total, total, total, total, total

    result_outer, _ = theano.scan(
        fn=outer_step,
        sequences=[a, b, c, d, e],
        name="outer_scan"
    )
```