@abergeron 2015-09-23
Overall I approve.  Would still have to look at the scan parts in detail before an ok though.