@carriepl 2015-09-25
I performed some more tests:

First, @nouiz wanted me to verify that this doesn't cause a significant increase in memory consumption. I tested on the LSTM tutorial from deeplearning.net and found a 300 KB increase in memory consumption on a total of 52MB according to Theano's profile_memory output. I think that difference is pretty trivial.

I also tested on one of the larger use cases provided by a student from the lab. Without this change, time to take the gradients was 30 minutes and compilation time was 2 hours. With the change, the time to take the gradients is 45s and compilation time is 1 hour 30 minutes. Total time to get the Theano function thus went fro 2h30 to 1h30.