@carriepl 2015-09-28
@nouiz : Still haven't been able to run his script. Will ask the student if he can spare the time to test the  speed up himself.

@abergeron or @lamblin : can one of you have a quick look at this?